feat: [G-2] uint24 for performanceFee

+++ contracts/vault/fee/PerfFeeAbstractVault.sol
@@ -25,12 +25,12 @@ import { FeeAdminAbstractVault } from "./FeeAdminAbstractVault.sol";
  */
 abstract contract PerfFeeAbstractVault is FeeAdminAbstractVault {
     /// @notice Scale of the performance fee. 100% = 1000000, 1% = 10000, 0.01% = 100
-    uint256 public constant FEE_SCALE = 1e6;
+    uint24 public constant FEE_SCALE = 1e6;
     /// @notice Scale of the assets per share used to calculate performance fees. 1e26 = 26 decimal places.
     uint256 public constant PERF_ASSETS_PER_SHARE_SCALE = 1e26;
 
     /// @notice Performance fee scaled to 6 decimal places. 1% = 10000, 0.01% = 100
-    uint256 public performanceFee;
+    uint24 public performanceFee;
 
     /// @notice Assets per shares used to calculate performance fees scaled to 26 decimal places.
     uint256 public perfFeesAssetPerShare;
@@ -39,7 +39,7 @@ abstract contract PerfFeeAbstractVault is FeeAdminAbstractVault {
     event PerformanceFeeUpdated(uint256 performanceFee);
 
     /// @param _performanceFee Performance fee scaled to 6 decimal places.
-    function _initialize(uint256 _performanceFee) internal virtual {
+    function _initialize(uint24 _performanceFee) internal virtual {
         require(_performanceFee <= FEE_SCALE, "Invalid fee");
         performanceFee = _performanceFee;
         perfFeesAssetPerShare = PERF_ASSETS_PER_SHARE_SCALE;
@@ -110,7 +110,7 @@ abstract contract PerfFeeAbstractVault is FeeAdminAbstractVault {
      * @notice Sets a new performance fee after charging to now using the old performance fee.
      * @param _performanceFee Performance fee scaled to 6 decimal places. 1% = 10000, 0.01% = 100
      */
-    function setPerformanceFee(uint256 _performanceFee) external onlyGovernor {
+    function setPerformanceFee(uint24 _performanceFee) external onlyGovernor {
         require(_performanceFee <= FEE_SCALE, "Invalid fee");
 
         // Charges a performance fee using the old value.
